% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklink.R
\name{checklink}
\alias{checklink}
\title{Set the Right Link Function for Specified Parameter and Distribution}
\usage{
checklink(which.link = NULL, which.dist = NULL, link = NULL, link.List = NULL)
}
\arguments{
\item{which.link}{which parameter link e.g. \code{which.link="mu.link"}}

\item{which.dist}{which distribution family e.g. \code{which.dist="Cole.Green"}}

\item{link}{a repetition of \code{which.link} e.g. \code{link=substitute(mu.link)}}

\item{link.List}{what link function are required e.g. \code{link.List=c("inverse", "log", "identity")}}
}
\value{
Defines the right link for each parameter.
}
\description{
This function is used within the distribution family specification of a
GAMLSS model to define the right link for each of the parameters of the
distribution. This function should not be called by the user unless
he/she specify a new distribution family or wishes to change existing
link functions in the parameters.
}
\details{
This function is the same \code{checklink} fuction from gamlss.dist package.
}
