% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d1_vec_dldm_compo_cpp}
\alias{d1_vec_dldm_compo_cpp}
\title{Function to obtain d1 vectorial in the score for COMPO with C++.}
\usage{
d1_vec_dldm_compo_cpp(mu, sigma)
}
\arguments{
\item{mu}{numeric vector.}

\item{sigma}{numeric vector.}
}
\description{
Function to obtain d1 vectorial in the score for COMPO with C++.
}
