% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH.R
\name{DBH}
\alias{DBH}
\title{The Discrete Burr Hatke family}
\usage{
DBH(mu.link = "logit")
}
\arguments{
\item{mu.link}{defines the mu.link, with "logit" link as the default for the mu parameter. Other links are "probit" and "cloglog"'(complementary log-log)}
}
\description{
The function \code{DBH()} defines the Discrete Burr Hatke distribution, one-parameter
discrete distribution, for a \code{gamlss.family} object to be used in GAMLSS fitting
using the function \code{gamlss()}.
}
\details{
The Discrete Burr-Hatke distribution with parameters \eqn{\mu} has a support
0, 1, 2, ... and density given by

\eqn{f(x | \mu) = (\frac{1}{x+1}-\frac{\mu}{x+2})\mu^{x}}

The pmf is log-convex for all values of \eqn{0 < \mu < 1}, where \eqn{\frac{f(x+1;\mu)}{f(x;\mu)}}
is an increasing function in \eqn{x} for all values of the parameter \eqn{\mu}.

Note: in this implementation we changed the original parameters \eqn{\lambda} for \eqn{\mu},
we did it to implement this distribution within gamlss framework.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
y <- rHYPERPO(n=300, mu=5, sigma=1.5)

# Fitting the model
library(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=HYPERPO,
               control=gamlss.control(n.cyc=500, trace=FALSE))

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what='mu'))
exp(coef(mod1, what='sigma'))

# Example 2
# Generating random values under some model

\dontrun{
  # A function to simulate a data set with Y ~ HYPERPO
  gendat <- function(n) {
    x1 <- runif(n)
    x2 <- runif(n)
    mu    <- exp(1.21 - 3 * x1) # 0.75 en promedio
    sigma <- exp(1.26 - 2 * x2) # 1.30 en promedio
    y <- rHYPERPO(n=n, mu=mu, sigma=sigma)
    data.frame(y=y, x1=x1,x2=x2)
  }

  set.seed(1235)
  datos <- gendat(n=150)

  mod2 <- NULL
  mod2 <- gamlss(y~x1, sigma.fo=~x2, family=HYPERPO, data=datos,
                 control=gamlss.control(n.cyc=500, trace=TRUE))

  summary(mod2)
}
}
\references{
\insertRef{el2020discrete}{DiscreteDists}
}
\seealso{
\link{dDBH}.
}
\author{
Valentina Hurtado Sepulveda, \email{vhurtados@unal.edu.co}
}
