% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dCOMPO_vec}
\alias{dCOMPO_vec}
\title{Function to obtain the dHYPERPO for a vector x}
\usage{
dCOMPO_vec(x, mu, sigma, log)
}
\arguments{
\item{x}{numeric value for x.}

\item{mu}{numeric value for mu.}

\item{sigma}{numeric value for sigma.}

\item{log}{logical value for log.}
}
\value{
returns the pmf for a vector.
}
\description{
Function to obtain the dHYPERPO for a vector x
}
\keyword{internal}
