% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dDGBH.R
\name{dDGBH}
\alias{dDGBH}
\alias{pDGBH}
\alias{qDGBH}
\alias{rDGBH}
\title{The Discrete Generalized Burr Hatke distribution}
\usage{
dDGBH(x, mu = 0.999, sigma = 100, log = FALSE)

pDGBH(q, mu = 0.999, sigma = 100, lower.tail = TRUE, log.p = FALSE)

qDGBH(p, mu = mu, sigma = sigma, lower.tail = TRUE)

rDGBH(n, mu = 0.999, sigma = 100)
}
\arguments{
\item{x, q}{vector of (non-negative integer) quantiles.}

\item{mu}{vector of the mu parameter.}

\item{sigma}{vector of the sigma parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X <= x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}
}
\value{
\code{dDGBH} gives the density, \code{pDGBH} gives the distribution
function, \code{qDGBH} gives the quantile function, \code{rDGBH}
generates random deviates.
}
\description{
These functions define the density, distribution function, quantile
function and random generation for the Discrete Generalized Burr Hatke distribution
with parameter \eqn{\mu}.
}
\details{
The Discrete Burr-Hatke distribution with parameters \eqn{\mu} has a support
0, 1, 2, ... and density given by

\eqn{f(x | \mu) = (\frac{1}{x+1}-\frac{\mu}{x+2})\mu^{x}}

The pmf is log-convex for all values of \eqn{0 < \mu < 1}, where \eqn{\frac{f(x+1;\mu)}{f(x;\mu)}}
is an increasing function in \eqn{x} for all values of the parameter \eqn{\mu}.

Note: in this implementation we changed the original parameters \eqn{\lambda} for \eqn{\mu},
we did it to implement this distribution within gamlss framework.
}
\examples{
# Example 1
dDGBH(x=0:10, mu=0.56, sigma=1.1)

# Example 2
# Checking if curves go to 1

# Example 3
# Some tests to compare the simulator with theoretical probabilities

# Example 4
}
\references{
\insertRef{yousof2021}{DiscreteDists}
}
\seealso{
\link{DGBH}.
}
\author{
Sofia Cuartas García, \email{scuartasg@unal.edu.co}
}
