% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGBH.R
\name{DGBH}
\alias{DGBH}
\title{The Discrete Generalized Burr Hatke family}
\usage{
DGBH(mu.link = "logit")
}
\arguments{
\item{mu.link}{defines the mu.link, with "logit" link as the default for the mu parameter. Other links are "probit" and "cloglog"'(complementary log-log)}
}
\description{
The function \code{DBH()} defines the Discrete Burr Hatke distribution, one-parameter
discrete distribution, for a \code{gamlss.family} object to be used in GAMLSS fitting
using the function \code{gamlss()}.
}
\details{
The Discrete Burr-Hatke distribution with parameters \eqn{\mu} has a support
0, 1, 2, ... and density given by

\eqn{f(x | \mu) = (\frac{1}{x+1}-\frac{\mu}{x+2})\mu^{x}}

The pmf is log-convex for all values of \eqn{0 < \mu < 1}, where \eqn{\frac{f(x+1;\mu)}{f(x;\mu)}}
is an increasing function in \eqn{x} for all values of the parameter \eqn{\mu}.

Note: in this implementation we changed the original parameters \eqn{\lambda} for \eqn{\mu},
we did it to implement this distribution within gamlss framework.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
y <- rDGBH(n=10, mu=0.65, sigma=1.5)
y

# Example 2
# Generating random values under some model

}
\references{
\insertRef{yousof2021}{DiscreteDists}
}
\seealso{
\link{dDGBH}.
}
\author{
Sofia Cuartas García, \email{scuartasg@unal.edu.co}
}
