% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEII.R
\name{DGEII}
\alias{DGEII}
\title{The DGEII family}
\usage{
DGEII(mu.link = "log", sigma.link = "logit")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "logit" link as the default for the sigma. Other links are "probit" and "cloglog"'(complementary log-log)}
}
\description{
The function \code{DGEII()} defines the Discrete generalized exponential distribution,
Second type, a two parameter
distribution, for a \code{gamlss.family} object to be used in GAMLSS fitting
using the function \code{gamlss()}.
}
\details{
The DGEII distribution with parameters \eqn{\mu} and \eqn{\sigma}
has a support 0, 1, 2, ... and mass function given by

\eqn{f(x | \mu, \sigma) = (1-\sigma^{x+1})^{\mu}-(1-\sigma^x)^{\mu}}

with \eqn{\mu > 0} and \eqn{0 < \sigma < 1}
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
set.seed(189)
y <- rDGEII(n=500, mu=10, sigma=0.25)

# Fitting the model
library(gamlss)
mod1 <- NULL
mod1 <- gamlss(y~1, family=DGEII,
               control=gamlss.control(n.cyc=500, trace=TRUE))

# Extracting the fitted values for mu and sigma
# using the inverse link function
inv_logit <- function(x) 1/(1 + exp(-x))
exp(coef(mod1, what="mu"))
inv_logit(coef(mod1, what="sigma"))


# Example 2
# Generating random values under some model

# A function to simulate a data set with Y ~ DGEII
gendat <- function(n) {
  x1 <- runif(n, min=0.10, max=0.80)
  x2 <- runif(n, min=0.10, max=0.80)
  mu    <- exp(1.7 - 2.8*x1)
  sigma <- inv_logit(0.73 - 3*x2)
  y <- rDGEII(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(654)

datos <- gendat(n=2000)

mod2 <- NULL
mod2 <- gamlss(y~x1, sigma.fo=~x2, family=DGEII, data=datos,
               control=gamlss.control(n.cyc=500, trace=TRUE))

summary(mod2)

}
\references{
\insertRef{nekoukhou2013}{DiscreteDists}
}
\seealso{
\link{dDGEII}.
}
\author{
Valentina Hurtado Sepúlveda, \email{vhurtados@unal.edu.co}
}
