% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estim_mu_DBH}
\alias{estim_mu_DBH}
\title{Initial values for Discrete Burr Hatke}
\usage{
estim_mu_DBH(y)
}
\arguments{
\item{y}{vector with the response variable.}
}
\value{
returns a scalar with the MLE estimation.
}
\description{
This function generates initial values for the parameter mu.
}
\keyword{internal}
