% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estim_mu_POISXL}
\alias{estim_mu_POISXL}
\title{Initial values for discrete Poisson XLindley distribution}
\usage{
estim_mu_POISXL(y)
}
\arguments{
\item{y}{vector with the response variable.}
}
\value{
returns a scalar with the MLE estimation.
}
\description{
This function generates initial values for the parameters.
}
\keyword{internal}
