% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPO.R
\name{estim_mu_sigma_COMPO}
\alias{estim_mu_sigma_COMPO}
\title{Initial values for COMPO}
\usage{
estim_mu_sigma_COMPO(y)
}
\arguments{
\item{y}{vector with the response variable.}
}
\value{
returns a vector with the moments estimations.
}
\description{
This function generates initial values for the parameters.
}
\keyword{internal}
