% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estim_mu_sigma_DGEII}
\alias{estim_mu_sigma_DGEII}
\title{Initial values for DGEII}
\usage{
estim_mu_sigma_DGEII(y)
}
\arguments{
\item{y}{vector with the response variable.}
}
\description{
This function generates initial values for the parameters.
}
\keyword{internal}
