% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.R
\name{estim_mu_sigma_DP}
\alias{estim_mu_sigma_DP}
\title{estim_mu_sigma_DP}
\usage{
estim_mu_sigma_DP(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
This function generates initial values for DP distribution.
}
\examples{
y <- rDP(n=100, mu=0.1, sigma=0.1)
estim_mu_sigma_DP(y=y)
}
