% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPERKS.R
\name{estim_mu_sigma_DPERKS}
\alias{estim_mu_sigma_DPERKS}
\title{Initial values for DPERKS}
\usage{
estim_mu_sigma_DPERKS(y)
}
\arguments{
\item{y}{vector with the response variable.}
}
\value{
returns a vector with the MLE estimations.
}
\description{
This function generates initial values for the parameters.
}
\keyword{internal}
