% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsPA.R
\name{estim_mu_sigma_DsPA}
\alias{estim_mu_sigma_DsPA}
\title{estim_mu_sigma_DsPA}
\usage{
estim_mu_sigma_DsPA(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
This function generates initial values for DsPA distribution
}
\examples{
y <- rDsPA(n=100, mu=0.5 , sigma=0.5)
estim_mu_sigma_DsPA(y=y)
}
