% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPGL.R
\name{estim_mu_sigma_NPGL}
\alias{estim_mu_sigma_NPGL}
\title{estim_mu_sigma_NPGL}
\usage{
estim_mu_sigma_NPGL(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
This function generates initial values for NPGL distribution.
}
\examples{
y <- rNPGL(n=100, mu=0.1, sigma=2)
estim_mu_sigma_NPGL(y=y)
}
