% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPGL2.R
\name{estim_mu_sigma_NPGL2}
\alias{estim_mu_sigma_NPGL2}
\title{estim_mu_sigma_NPGL2}
\usage{
estim_mu_sigma_NPGL2(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
estim_mu_sigma_NPGL2
}
\examples{
y <- rNPGL2(n = 100, mu = 3, sigma = 2)
estim_mu_sigma_NPGL2(y = y)
}
