% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPGL.R
\name{estim_mu_sigma_NPGL_MM}
\alias{estim_mu_sigma_NPGL_MM}
\title{estim_mu_sigma_NPGL_MM}
\usage{
estim_mu_sigma_NPGL_MM(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
This function generates initial values for NPGL distribution
using the method of moments estimators given in equations (19) and (20)
of Altun (2021).
}
\examples{
y <- rNPGL(n=100, mu=0.1, sigma=2)
estim_mu_sigma_NPGL_MM(y=y)
}
