% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grazing.R
\docType{data}
\name{grazing}
\alias{grazing}
\title{grazing dataset}
\format{
\subsection{\code{grazing}}{

A data frame with 62 rows and 3 variables:
\describe{
\item{when}{when the bird count was conducted}
\item{grazed}{a factor with levels Reference and Feral}
\item{birds}{the number of understorey birds}
...
}
}
}
\source{
\url{https://github.com/rdmatheus/bergreg}
}
\usage{
grazing
}
\description{
In this experiment, the density of understorey birds at
a series of sites in two areas either side of a stockproof fence
were compared. Once side had limited grazing (mainly from native
herbivores), and the other was heavily grazed by feral herbivores,
mostly horses. Bird counts were done at the sites either side of
the fence (the Before measurements). Then the herbivores were
removed, and bird counts done again (the After measurements).
The measurements are the total number of understorey-foraging
birds observed in three 20-minute surveys of two hectare quadrats.
}
\author{
Rodrigo Matheus
}
\keyword{datasets}
