% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik_DGEII}
\alias{logLik_DGEII}
\title{logLik function for DGEII}
\usage{
logLik_DGEII(transf_param = c(0, 0), x)
}
\arguments{
\item{x}{vector with the response variable.}

\item{param}{vector with parameters in log and logit scale.}
}
\description{
Calculates logLik for DGEII distribution.
}
\keyword{internal}
