% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik_DIKUM}
\alias{logLik_DIKUM}
\title{logLik function for discrete Inverted Kumaraswamy}
\usage{
logLik_DIKUM(param = c(0, 0), x)
}
\arguments{
\item{param}{vector with parameters in log scale.}

\item{x}{vector with the response variable.}
}
\value{
returns the loglikelihood given the parameters and random sample.
}
\description{
Calculates logLik for discrete Inverted Kumaraswamy distribution.
}
\keyword{internal}
