% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik_DMOLBE}
\alias{logLik_DMOLBE}
\title{logLik function for DMOLBE}
\usage{
logLik_DMOLBE(logparam = c(0, 0), x)
}
\arguments{
\item{logparam}{vector with parameters in log scale.}

\item{x}{vector with the response variable.}
}
\value{
returns the loglikelihood given the parameters and random sample.
}
\description{
Calculates logLik for DMOLBE distribution.
}
\keyword{internal}
