% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.R
\name{logLik_DP}
\alias{logLik_DP}
\title{logLik_DP}
\usage{
logLik_DP(logparam = c(0, 0), x)
}
\arguments{
\item{logparam}{vector with the values for mu sigma}

\item{x}{vector with the data}
}
\description{
This is an auxiliar function to obtain the logLik for DP.
}
\examples{
y <- rDP(n=100, mu=0.5, sigma=0.5)
logLik_DP(logparam=c(0, 0), x=y)
}
