% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPERKS.R
\name{logLik_DPERKS}
\alias{logLik_DPERKS}
\title{logLik_DPERKS}
\usage{
logLik_DPERKS(logparam = c(0, 0), x)
}
\arguments{
\item{logparam}{vector with the values for mu sigma}

\item{x}{vector with the data}
}
\description{
This is an auxiliar function to obtain the logLik for DPERKS.
}
\examples{
y <- rDPERKS(n=100, mu=0.5, sigma=0.5)
logLik_DPERKS(logparam=c(0, 0), x=y)
}
