% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsPA.R
\name{logLik_DsPA}
\alias{logLik_DsPA}
\title{logLik function for DsPA}
\usage{
logLik_DsPA(param = c(0, 0), x)
}
\arguments{
\item{param}{a vector with parameters.}

\item{x}{vector with the response variable.}
}
\value{
returns the loglikelihood given the parameters and random sample.
}
\description{
Calculates logLik for DsPA distribution.
}
\keyword{internal}
