% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_mu_sigma_HYPERPO.R
\name{logLik_HYPERPO}
\alias{logLik_HYPERPO}
\title{logLik function for hyper Poisson}
\usage{
logLik_HYPERPO(logparam = c(0, 0), x)
}
\arguments{
\item{logparam}{vector with parameters in log scale.}

\item{x}{vector with the response variable.}
}
\description{
Calculates logLik for hyper Poisson distribution.
}
\keyword{internal}
