% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik_HYPERPO2}
\alias{logLik_HYPERPO2}
\title{logLik function for hyper Poisson in second parameterization}
\usage{
logLik_HYPERPO2(logparam = c(0, 0), x)
}
\arguments{
\item{logparam}{vector with parameters in log scale.}

\item{x}{vector with the response variable.}
}
\value{
returns the loglikelihood given the parameters and random sample.
}
\description{
Calculates logLik for hyper Poisson distribution.
}
\keyword{internal}
