% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPGL.R
\name{logLik_NPGL}
\alias{logLik_NPGL}
\title{logLik_NPGL}
\usage{
logLik_NPGL(param = c(0, 0), x)
}
\arguments{
\item{param}{vector with the values for mu and sigma}

\item{x}{vector with the data}
}
\description{
This is an auxiliar function to obtain the logLik for NPGL.
}
\examples{
y <- rNPGL(n=100, mu=0.1, sigma=2)
logLik_NPGL(param=c(0, 0), x=y)
}
