% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPGL2.R
\name{logLik_NPGL2}
\alias{logLik_NPGL2}
\title{logLik_NPGL2}
\usage{
logLik_NPGL2(param = c(log(3), log(2)), x)
}
\arguments{
\item{param}{vector with the values for mu and sigma}

\item{x}{vector with the data}
}
\description{
Auxiliary function to evaluate the log-likelihood of the mean-parametrized
NPGL2 distribution.
}
\examples{
y <- rNPGL2(n = 100, mu = 3, sigma = 2)
logLik_NPGL2(param = c(log(3), log(2)), x = y)
}
