% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.link.gamlss.R
\name{make.link.gamlss}
\alias{make.link.gamlss}
\title{Create a Link for GAMLSS families}
\usage{
make.link.gamlss(link)
}
\arguments{
\item{link}{character or numeric; one of \code{"logit"}, \code{"probit"},
\code{"cloglog"}, \code{"identity"}, \code{"log"},  \code{"sqrt"},
\code{"1/mu^2"}, \code{"inverse"}, \code{"logshifted"},
\code{"logitshifted"}, or number, say lambda resulting in power link
\eqn{\mu^\lambda}{mu^lambda}.}
}
\value{
For the \code{make.link.gamlss} a list with components

linkfun: Link function \code{function(parameter)}

linkinv: Inverse link function \code{function(eta)}

mu.eta: Derivative \code{function(eta)} dparameter/deta

valideta: \code{function(eta)} TRUE if all of eta is in the domain of \code{linkinv}.

For the \code{show.link} a list with components the available links for the distribution parameters
}
\description{
The function \code{make.link.gamlss()} is used with \code{gamlss.family}
distributions in package \pkg{gamlss()}.Given a link, it returns a link
function, an inverse link  function, the derivative dpar/deta where 'par'
is the appropriate distribution parameter and a function for checking
the domain. It differs from the usual \code{make.link} of \code{glm()} by
having extra links as the \code{logshifto1}, and the \code{own}.
For the use of the \code{own} link see the example bellow.
\code{show.link} provides a way in which the user can identify the link
functions available for each gamlss distribution. If your required link
function is not available for any of the gamlss distributions you can add it in.
}
\details{
The \code{own} link function is added to allow the user greater
flexibility. In order to used the own link function for any of the
parameters of the distribution the \code{own} link should  appear in the
available links for this parameter. You can check this using the function
\code{show.link}. If the \code{own} do not appear in the list you can
create a new function for the distribution in which \code{own} is added
in the list. For example the first line of the code  of the binomial
distribution, \code{BI}, has change from

"mstats <- checklink("mu.link", "Binomial", substitute(mu.link),
c("logit", "probit", "cloglog", "log")), in version 1.0-0 of gamlss, to

"mstats <- checklink("mu.link", "Binomial", substitute(mu.link),
c("logit", "probit", "cloglog", "log", "own"))

in version 1.0-1. Given that the parameter has \code{own} as an option the user needs also
to define the following four new functions in order to used an
\code{own} link.

i) own.linkfun

ii) own.linkinv

iii) own.mu.eta and

iv) own.valideta.

An example is given below.

Only one parameter of the distribution at a time is allowed to have
its \code{own} link, (unless the same four \code{own} functions above
are suitable for more that one parameter of the distribution).

Note that from \pkg{gamlss} version 1.9-0 the user can introduce its own
link function by define an appropriate function, (see the example below).
}
\note{
For the links  involving parameters as in \code{logshifted} and
\code{logitshifted} the parameters can be passed in the definition of the
distribution by calling the  \code{ checklink} function, for example in the
definition of the \code{tau} parameter in BCPE distribution the following
call is made: \code{tstats <- checklink("tau.link",
"Box Cox Power Exponential", substitute(tau.link),
c("logshifted", "log", "identity"), par.link = c(1))  }

This function is the same \code{make.link.gamlss} fuction from gamlss.dist package.
}
