% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMES1988.R
\docType{data}
\name{NMES1988}
\alias{NMES1988}
\title{NMES1988 dataset}
\format{
\subsection{\code{NMES1988}}{

A data frame containing 4,406 observations on 19 variables:
\describe{
\item{visits}{Number of physician office visits.}
\item{nvisits}{Number of non-physician office visits.}
\item{ovisits}{Number of physician hospital outpatient visits.}
\item{novisits}{Number of non-physician hospital outpatient visits.}
\item{emergency}{Emergency room visits.}
\item{hospital}{Number of hospital stays.}
\item{health}{Factor indicating self-perceived health status, levels are "poor", "average" (reference category), "excellent".}
\item{chronic}{Number of chronic conditions.}
\item{adl}{Factor indicating whether the individual has a condition that limits activities of daily living ("limited") or not ("normal").}
\item{region}{Factor indicating region, levels are northeast, midwest, west, other (reference category).}
\item{age}{Age in years (divided by 10).}
\item{afam}{Factor. Is the individual African-American?}
\item{gender}{Factor indicating gender.}
\item{married}{Factor. Is the individual married?}
\item{school}{Number of years of education.}
\item{income}{Family income in USD 10,000.}
\item{employed}{Factor. Is the individual employed?}
\item{insurance}{Factor. Is the individual covered by private insurance?}
\item{medicaid}{Factor. Is the individual covered by Medicaid?}
}
}
}
\source{
This dataset comes from the AER R package.
}
\usage{
NMES1988
}
\description{
Cross-section data originating from the US National Medical Expenditure
Survey (NMES) conducted in 1987 and 1988. The NMES is based upon a
representative, national probability sample of the civilian
non-institutionalized population and individuals admitted to long-term
care facilities during 1987. The data are a subsample of individuals ages
66 and over all of whom are covered by Medicare (a public insurance program
providing substantial protection against health-care costs).
}
\author{
AER R package.
}
\keyword{datasets}
