% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPGL.R
\name{NPGL}
\alias{NPGL}
\title{New Poisson-generalised Lindley distribution}
\usage{
NPGL(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}
}
\value{
Returns a \code{gamlss.family} object which can be used
to fit a NPGL distribution
in the \code{gamlss()} function.
}
\description{
The function \code{NPGL()} defines the
Poisson-generalised Lindley distribution,
a two parameter distribution,
for a \code{gamlss.family} object to be used in GAMLSS
fitting using the function \code{gamlss()}.
}
\details{
The Poisson-generalised Lindley distribution with parameters \eqn{\mu} and \eqn{\sigma}
has support \eqn{x = 0, 1, 2, \ldots} and probability mass function given by

\eqn{f(x \mid \mu, \sigma)=\frac{\mu^2+\frac{\mu^{\sigma}(\mu+1)^{1-\sigma}\Gamma(x+\sigma)}{\Gamma(\sigma)\Gamma(x+1)}}{(\mu+1)^{x+2}}}

with \eqn{\mu > 0} and \eqn{\sigma > 0}.

This distribution is useful for modeling over-dispersed count data.

Note: in this implementation we changed the original parameters \eqn{\theta} and \eqn{\alpha}
for \eqn{\mu} and \eqn{\sigma} respectively, we did it to implement this distribution within gamlss framework.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma

set.seed(123)
y <- rNPGL(n=100, mu=20, sigma=2)

# Fitting the model
library(gamlss)
mod1 <- gamlss(y~1, family=NPGL,
               control=gamlss.control(n.cyc=500, trace=FALSE))

# Extracting the fitted values for mu and sigma
exp(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))

# Example 2
# Generating random values under some model

# A function to simulate a data set with Y ~ NPGL
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu    <- exp(1.7 - 2.8 * x1) # Approx 1.35
  sigma <- exp(0.73 + 1 * x2)  # Approx 3.42
  y <- rNPGL(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(1234)
datos <- gendat(n=200)

mod2 <- gamlss(y~x1, sigma.fo=~x2, family=NPGL, data=datos,
               control=gamlss.control(n.cyc=800, trace=FALSE))

summary(mod2)
}
\references{
Altun, E.
A new two-parameter discrete poisson-generalized Lindley distribution with
properties and applications to healthcare data sets. Comput Stat 36, 2841–2861 (2021). https://doi.org/10.1007/s00180-021-01097-0
}
\seealso{
\link{dNPGL}.
}
\author{
Tomás Mesa, \email{tomas.mesaz@udea.edu.co}
}
