% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtaining_lambda}
\alias{obtaining_lambda}
\title{Auxiliar function to obtain lambda from E(X)}
\usage{
obtaining_lambda(media, gamma)
}
\arguments{
\item{media}{the value for the mean or E(X).}

\item{gamma}{the value for the gamma parameter.}
}
\value{
returns the value of lambda to ensure the mean and gamma.
}
\description{
This function implements the procedure given in page 150.
}
\keyword{internal}
