% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POISXL.R
\name{POISXL}
\alias{POISXL}
\title{The Discrete Poisson XLindley}
\usage{
POISXL(mu.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}
}
\description{
The function \code{POISXL()} defines  the Discrete Poisson XLindley distribution, one-parameter
discrete distribution, for a \code{gamlss.family} object to be used in GAMLSS fitting
using the function \code{gamlss()}.
}
\details{
The Discrete Poisson XLindley distribution with parameters \eqn{\mu} has a support
0, 1, 2, ... and mass function given by

\eqn{f(x | \mu) = \frac{\mu^2(x+\mu^2+3(1+\mu))}{(1+\mu)^{4+x}}}; with \eqn{\mu>0}
Note: in this implementation we changed the original parameters \eqn{\theta} for \eqn{\mu},
we did it to implement this distribution within gamlss framework.
}
\examples{
# Example 1
# Generating some random values with
# known mu
y <- rPOISXL(n=1000, mu=1)

estim_mu_POISXL(y)

# Fitting the model
library(gamlss)
mod1 <- gamlss(y~1, family=POISXL,
               control=gamlss.control(n.cyc=500, trace=FALSE))

# Extracting the fitted values for mu
# using the inverse link function
exp(coef(mod1, what='mu'))

}
\references{
\insertRef{ahsan2022}{DiscreteDists}
}
\seealso{
\link{dPOISXL}.
}
\author{
Mariana Blandon Mejia, \email{mblandonm@unal.edu.co}
}
