% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_hp}
\alias{simulate_hp}
\title{Auxiliar function to generate values for hyper Poisson}
\usage{
simulate_hp(sigma, mu)
}
\arguments{
\item{sigma, mu}{values to simulate.}
}
\description{
This function is used inside random function of Hyper Poisson.
}
\keyword{internal}
