% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{z_cpp}
\alias{z_cpp}
\title{Function to obtain Z for COMPO with C++.}
\usage{
z_cpp(lambda, nu, max_terms = 1000L, tol = 1e-10)
}
\arguments{
\item{lambda}{numeric value for mu.}

\item{nu}{numeric value for sigma.}

\item{max_terms}{numeric value.}

\item{tol}{numeric value.}
}
\value{
returns the z value.
}
\description{
Function to obtain Z for COMPO with C++.
}
\keyword{internal}
