% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{z_vec_cpp}
\alias{z_vec_cpp}
\title{Function to obtain Z vectorial for COMPO with C++.}
\usage{
z_vec_cpp(mu, sigma)
}
\arguments{
\item{mu}{numeric vector.}

\item{sigma}{numeric vector.}
}
\value{
returns the z value.
}
\description{
Function to obtain Z vectorial for COMPO with C++.
}
\keyword{internal}
