% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AR}
\alias{AR}
\title{Auxiliar function for hyper Poisson}
\usage{
AR(a, r)
}
\arguments{
\item{a}{first value.}

\item{r}{second value.}
}
\value{
returns the value for the a(r) function.
}
\description{
This function is used to calculate (a)r.
}
\keyword{internal}
