% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPERKS.R
\name{estim_mu_sigma_DPERKS}
\alias{estim_mu_sigma_DPERKS}
\title{estim_mu_sigma_DPERKS}
\usage{
estim_mu_sigma_DPERKS(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
This function generates initial values for DPERKS distribution.
}
\examples{
y <- rDPERKS(n=100, mu=0.1, sigma=0.1)
estim_mu_sigma_DPERKS(y=y)
}
