% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsPA.R
\name{estim_mu_sigma_DsPA}
\alias{estim_mu_sigma_DsPA}
\title{Initial values for DsPA}
\usage{
estim_mu_sigma_DsPA(y)
}
\arguments{
\item{y}{vector with the response variable.}
}
\value{
returns a vector with the MLE estimations.
}
\description{
This function generates initial values for the parameters.
}
\keyword{internal}
