% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estim_mu_sigma_HYPERPO}
\alias{estim_mu_sigma_HYPERPO}
\title{Initial values for hyper Poisson}
\usage{
estim_mu_sigma_HYPERPO(y)
}
\arguments{
\item{y}{vector with the response variable.}
}
\value{
returns a vector with the MLE estimations.
}
\description{
This function generates initial values for the parameters.
}
\keyword{internal}
