% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{F11}
\alias{F11}
\title{Auxiliar function for hyper Poisson}
\usage{
F11(a, c, z)
}
\arguments{
\item{a, c, z}{values for F11.}
}
\description{
This function is used inside density function of Hyper Poisson.
}
\keyword{internal}
