% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik_DLD}
\alias{logLik_DLD}
\title{logLik function for Discrete Lindley distribution}
\usage{
logLik_DLD(param = 0.5, x)
}
\arguments{
\item{param}{value for mu.}

\item{x}{vector with the response variable.}
}
\value{
returns the loglikelihood given the parameters and random sample.
}
\description{
Calculates logLik for Discrete Lindley  distribution.
}
\keyword{internal}
