% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPERKS.R
\name{logLik_DPERKS}
\alias{logLik_DPERKS}
\title{logLik function for DPERKS}
\usage{
logLik_DPERKS(logparam = c(0, 0), x)
}
\arguments{
\item{logparam}{vector with parameters in log scale.}

\item{x}{vector with the response variable.}
}
\value{
returns the loglikelihood given the parameters and random sample.
}
\description{
Calculates logLik for DPERKS distribution.
}
\keyword{internal}
