% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DsPA.R
\name{logLik_DsPA}
\alias{logLik_DsPA}
\title{logLik_DsPA
This is an auxiliar function to obtain the logLik for DsPA}
\usage{
logLik_DsPA(param = c(0, 0), x)
}
\arguments{
\item{param}{vector with the values for mu and sigma}

\item{x}{vector with the data}
}
\description{
logLik_DsPA
This is an auxiliar function to obtain the logLik for DsPA
}
\examples{
y <- rDsPA(n=100, mu=0.5 , sigma=0.5)
logLik_DsPA(param=c(0, 0), x=y)
}
