% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik_POISXL}
\alias{logLik_POISXL}
\title{logLik function for Poisson XLindley distribution}
\usage{
logLik_POISXL(param = 0, x)
}
\arguments{
\item{param}{parameter mu in log scale.}

\item{x}{vector with the response variable.}
}
\value{
returns the loglikelihood given the parameters and random sample.
}
\description{
Calculates logLik for Poisson XLindley distribution distribution.
}
\keyword{internal}
