% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hp.R, R/utils.R
\name{simulate_hp}
\alias{simulate_hp}
\title{The simulate_hp}
\usage{
simulate_hp(sigma, mu)

simulate_hp(sigma, mu)
}
\arguments{
\item{sigma}{value for sigma parameter.}

\item{mu}{value for mu parameter.}
}
\value{
a single value for the HYPERPO distribution.
}
\description{
Auxiliar function to generate a single observation for HYPERPO.

This function is used inside random function of Hyper Poisson.
}
\keyword{internal}
