% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stopping}
\alias{stopping}
\title{Auxiliar function for F11}
\usage{
stopping(x, tol)
}
\arguments{
\item{x}{vector}

\item{tol}{this is the tolerance of the infinite sum.}
}
\value{
returns a logical value if the tolerance level is met.
}
\description{
This function is used inside F11 function.
}
\keyword{internal}
